+++ tests/test-clients/cpp-libs/sharedlib.cpp
@@ -22,7 +22,7 @@ double SharedLib::bar(unsigned long max) const
     std::uniform_real_distribution<double> uniform(-1E5, 1E5);
     std::default_random_engine engine;
     double s = 0;
-    for (int i = 0; i < max; ++i) {
+    for (unsigned long i = 0; i < max; ++i) {
         s += asdf(uniform(engine), uniform(engine));
     }
     return s;
